+++ Model/Menu/NodeRepository.php
@@ -28,14 +28,21 @@ class NodeRepository implements NodeRepositoryInterface
      */
     protected $collectionFactory;
 
+    /**
+     * @var Cache
+     */
+    private $cache;
+
     public function __construct(
         NodeFactory $objectFactory,
         CollectionFactory $collectionFactory,
-        SearchResultsInterfaceFactory $searchResultsFactory
+        SearchResultsInterfaceFactory $searchResultsFactory,
+        Cache $cache
     ) {
         $this->objectFactory = $objectFactory;
         $this->collectionFactory = $collectionFactory;
         $this->searchResultsFactory = $searchResultsFactory;
+        $this->cache = $cache;
     }
 
     /**
@@ -44,7 +51,9 @@ class NodeRepository implements NodeRepositoryInterface
     public function save(NodeInterface $object)
     {
         try {
+            $hasObjectDataChanged = $object->hasDataChanges();
             $object->save();
+            $this->cache->invalidatePageCache($hasObjectDataChanged);
         } catch (Exception $e) {
             throw new CouldNotSaveException($e->getMessage());
         }