+++ Model/MenuRepository.php
@@ -26,21 +26,27 @@ class MenuRepository implements MenuRepositoryInterface
     /** @var ResourceModel\Menu */
     private $menuResourceModel;
 
+    /** @var Cache */
+    private $cache;
+
     /**
      * @param MenuFactory $menuFactory
      * @param CollectionFactory $menuCollectionFactory
      * @param MenuSearchResultsInterfaceFactory $menuSearchResults
+     * @param Menu\Cache $cache
      * @param ResourceModel\Menu|null $menuResourceModel
      */
     public function __construct(
         MenuFactory $menuFactory,
         CollectionFactory $menuCollectionFactory,
         MenuSearchResultsInterfaceFactory $menuSearchResults,
+        Menu\Cache $cache,
         ResourceModel\Menu $menuResourceModel = null
     ) {
         $this->menuFactory = $menuFactory;
         $this->collectionFactory = $menuCollectionFactory;
         $this->menuSearchResultsFactory = $menuSearchResults;
+        $this->cache = $cache;
         $this->menuResourceModel = $menuResourceModel
             ?? ObjectManager::getInstance()->get(ResourceModel\Menu::class); // Backwards-compatible class loader
     }
@@ -51,7 +57,9 @@ class MenuRepository implements MenuRepositoryInterface
     public function save(MenuInterface $menu)
     {
         try {
+            $hasObjectDataChanged = $object->hasDataChanges();
             $this->menuResourceModel->save($menu);
+            $this->cache->invalidatePageCache($hasObjectDataChanged);
         } catch (\Exception $e) {
             throw new CouldNotSaveException($e->getMessage());
         }