+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Functionality to enable/disable specific menu node (#85, DEV-65561)
+- Invalidate page cache on menu save (#70191)
 
 ## [2.12.0] - 2020-08-24
 ### Added